Refactor generator logic
