+++ src/cosi_db/controller/person.rs
@@ -11,8 +11,9 @@ use rocket::futures::TryStreamExt;
 // cosi_db
 use crate::cosi_db::connection::{CosiDB, MongoConnection};
 use crate::cosi_db::controller::common::{get_connection, PaginateData};
-use crate::cosi_db::generator::Generator;
+use crate::cosi_db::model::common::Generator;
 use crate::cosi_db::model::person::Person;
+
 use crate::{generate_generators, generate_pageable_getter};
 
 generate_generators! { Person, "person" }
