+++ src/cosi_db/mod.rs
@@ -1,4 +1,3 @@
 pub mod connection;
 pub mod controller;
-pub mod generator;
 pub mod model;
