+++ src/cosi_db/model/common.rs
@@ -1,0 +1,12 @@
+use async_trait::async_trait;
+use mongodb::Collection;
+
+#[async_trait]
+pub trait Generator<T> {
+    fn generate(size: u32) -> Vec<T>;
+}
+
+#[async_trait]
+pub trait COSICollection<T> {
+    async fn get_collection() -> Collection<T>;
+}
