+++ src/cosi_db/model/mod.rs
@@ -1,3 +1,4 @@
 pub mod address;
+pub mod common;
 pub mod household;
 pub mod person;
